FIREFLY-1256: Move PROP_SHEET declaration to PropertySheet.jsx
