+++ src/firefly/js/templates/fireflyviewer/TriViewPanel.jsx
@@ -22,9 +22,8 @@ import {getExpandedChartProps} from '../../charts/ChartsCntlr.js';
 import {DEFAULT_PLOT2D_VIEWER_ID} from '../../visualize/MultiViewCntlr.js';
 import {usePinnedChartInfo, PinnedChartPanel, PINNED_VIEWER_ID, BadgeLabel} from 'firefly/charts/ui/PinnedChartContainer.jsx';
 import {allowPinnedCharts} from '../../charts/ChartUtil.js';
-import {PropertySheetAsTable} from 'firefly/tables/ui/PropertySheet';
+import {PropertySheetAsTable, PROP_SHEET} from 'firefly/tables/ui/PropertySheet';
 import {getAppOptions} from 'firefly/core/AppDataCntlr';
-import {PROP_SHEET} from 'firefly/tables/TableUtil';
 
 const stateKeys= ['title', 'mode', 'showTables', 'showImages', 'showXyPlots', 'images'];
 const LEFT= 'LEFT';
